lobby fix new

+++ srcs/trascendence/public/website/js/utils_main/error_main.js
@@ -1,6 +1,6 @@
 import { free_users } from "../security/security.js";
 import { resetMatchStatsData } from "../game/pong/data/game_stats.js";
-import { navigate, reset_all_let, token, current_user, opponent, save_global, in_game, user_name, acess, invalid } from "../main.js";
+import { navigate, reset_all_let, token, current_user, opponent, save_global, in_game, acess, invalid } from "../main.js";
 import { showInfoModal } from "../modal.js";
 import { resetBracketState } from "../pages/tournament/bracket.js";
 import { addCallbackPageHandlers } from "../login/login_logic.js";
@@ -43,7 +43,6 @@ async function not_home(path) {
     }
     if (!opponent && (path === "/classic" || path === "/forza4/game")) {
         navigate("/modes", "return to modes");
-        console.log("sos");
         showInfoModal("the operation you are doing is forbidden", () => { });
         return (1);
     }
@@ -58,7 +57,7 @@ async function not_home(path) {
         resetBracketState();
         resetMatchStatsData();
     }
-    if (acess === false) {
+    if (acess === false && !token) {
         showInfoModal("ERROR: accessing unauthorized page...", () => { });
         await remove_all();
         return (1);
